[TASK] Bring back compabibility with TYPO3 v11
There is no reasons to migrate this extension and only make it
compatible with TYPO3 v12 since it was previously compatible
with both TYPO3 v10 and v11.

As TYPO3 v10 may now be considered obsolete and since there is
an official release supporting both TYPO3 v10 and v11, we choose
to further keep compatibility with TYPO3 v11 and v12 only.